Разделение контроллера склада на два и разделение страниц склада. + reformat code
