+++ src/main/java/ru/wms/WarehouseManagementService/controller/EmployeeController.java
@@ -5,10 +5,8 @@ import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import ru.wms.WarehouseManagementService.dto.UserRegistrationDTO;
 import ru.wms.WarehouseManagementService.entity.Employee;
 import ru.wms.WarehouseManagementService.security.UserPrincipal;
 import ru.wms.WarehouseManagementService.service.EmployeeService;
@@ -26,6 +24,7 @@ public class EmployeeController {
     public String getEmployee(@AuthenticationPrincipal UserPrincipal userPrincipal, Model model) {
         model.addAttribute("employees", employeeService.findMyEmployee(userPrincipal.getUser()));
         model.addAttribute("employee", new Employee());
+
         return "/employees";
     }
