+++ src/main/java/ru/wms/WarehouseManagementService/entity/Employee.java
@@ -24,7 +24,6 @@ public class Employee {
 
     private String patronymic;
 
-
     @OneToOne
     @JoinColumn(name = "user_id", nullable = false)
     private User user;
