+++ src/main/java/ru/wms/WarehouseManagementService/repository/CustomerRepository.java
@@ -3,9 +3,6 @@ package ru.wms.WarehouseManagementService.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ru.wms.WarehouseManagementService.entity.Customer;
-import ru.wms.WarehouseManagementService.entity.User;
-
-import java.util.Optional;
 
 @Repository
 public interface CustomerRepository extends JpaRepository<Customer, Long> {
