+++ src/main/java/ru/wms/WarehouseManagementService/repository/ProductRepository.java
@@ -1,6 +1,5 @@
 package ru.wms.WarehouseManagementService.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 import ru.wms.WarehouseManagementService.entity.Product;
