+++ src/main/java/ru/wms/WarehouseManagementService/security/Authority.java
@@ -12,6 +12,7 @@ public enum Authority implements GrantedAuthority {
 
     @Override
     public String getAuthority() {
+
         return name();
     }
 }
