+++ src/main/java/ru/wms/WarehouseManagementService/security/UserPrincipal.java
@@ -18,39 +18,45 @@ public class UserPrincipal implements UserDetails {
         this.user = user;
     }
 
-
     @Override
     public Collection<? extends GrantedAuthority> getAuthorities() {
+
         return this.user.getAuthorities();
     }
 
     @Override
     public String getPassword() {
+
         return this.user.getPassword();
     }
 
     @Override
     public String getUsername() {
+
         return this.user.getEmail();
     }
 
     @Override
     public boolean isAccountNonExpired() {
+
         return true;
     }
 
     @Override
     public boolean isAccountNonLocked() {
+
         return true;
     }
 
     @Override
     public boolean isCredentialsNonExpired() {
+
         return true;
     }
 
     @Override
     public boolean isEnabled() {
+
         return this.user.isActive();
     }
