+++ src/main/resources/templates/index.html
@@ -22,7 +22,7 @@
 
     <a href="/products">Товары</a>  <br>
 
-    <a href="/orders">Создать заказ</a> <br>
+    <a href="/orders">Заказы</a> <br>
 
     <a href="/employees" sec:authorize="hasRole('CUSTOMER')">Создать работника</a> <br>
