+++ src/main/resources/templates/warehouse/createWarehouse.html
@@ -1,0 +1,59 @@
+<!DOCTYPE html>
+<html lang="ru" xmlns:th="http://www.thymeleaf.org"
+      xmlns:sec="http://www.thymeleaf.org/extras/spring-security">
+<head>
+    <meta charset="UTF-8">
+    <title>Warehouse</title>
+</head>
+<body>
+
+<div class="container">
+
+    <h2>Вы вошли в систему как <span sec:authentication="name"></span></h2>
+
+    <hr>
+        <p>
+        Раздел управления складами (регистрация склада)
+        </p>
+    <hr>
+
+    <hr>
+        <h2>Создать склад</h2>
+    <hr>
+    <form class="mb-3" th:action="@{/createWarehouse}" th:object="${warehouse}" method="post">
+        <div class="form-group mb-3">
+            <label for="name">Имя склада</label>
+            <input type="text" class="form-control" id="name" placeholder="имя склада" value="name" th:field="*{name}" required>
+        </div>
+        <div class="form-group mb-3">
+            <label for="address">Адрес</label>
+            <input type="text" class="form-control" id="address" placeholder="адрес" value="address" th:field="*{address}" required>
+        </div>
+        <div class="form-group mb-3">
+            <label for="capacity">Площадь</label>
+            <input type="number" class="form-control" id="capacity" placeholder="123" value="capacity" th:field="*{capacity}" required>
+        </div>
+        <div class="form-group mb-3">
+            <label for="contactInformation">Контактная информация</label>
+            <input type="text" class="form-control" id="contactInformation" placeholder="контактная информация" value="contact information" th:field="*{contactInformation}" required>
+        </div>
+        <div class="form-group mb-3 ">
+            <label for="creationDate">Дата создания</label>
+            <input type="date" class="form-control" id="creationDate" placeholder="01.01.2000" value="01.01.2001" th:field="*{creationDate}" required>
+        </div>
+
+        <input type="submit" value="Зарегистрировать склад" />
+        <a href="/warehouses">Отменить регистрацию склада</a>
+    </form>
+    <hr>
+
+    <h3><a href="/warehouses">Вернутся в главное меню склада</a></h3>
+
+    <h3><a href="/index">Вернутся в основное меню</a></h3>
+
+</div>
+
+<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css">
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
+</body>
+</html>
