+++ src/main/resources/templates/warehouse/warehouses.html
@@ -12,41 +12,22 @@
     <h2>Вы вошли в систему как <span sec:authentication="name"></span></h2>
 
     <hr>
-    <p>
-        Раздел управления складами
-    </p>
+        <p>
+         Раздел управления складами
+        </p>
     <hr>
 
-    <h3><a href="/index">Вернуться в меню</a></h3>
-    <hr>
-    <h2>Создать склад</h2>
-    <hr>
-    <form class="mb-3" th:action="@{/warehouses}" th:object="${warehouse}" method="post">
-        <div class="form-group mb-3">
-            <label for="name">Имя склада</label>
-            <input type="text" class="form-control" id="name" placeholder="имя склада" value="name" th:field="*{name}" required>
-        </div>
-        <div class="form-group mb-3">
-            <label for="address">Адрес</label>
-            <input type="text" class="form-control" id="address" placeholder="адрес" value="address" th:field="*{address}" required>
-        </div>
-        <div class="form-group mb-3">
-            <label for="capacity">Площадь</label>
-            <input type="number" class="form-control" id="capacity" placeholder="123" value="capacity" th:field="*{capacity}" required>
-        </div>
-        <div class="form-group mb-3">
-            <label for="contactInformation">Контактная информация</label>
-            <input type="text" class="form-control" id="contactInformation" placeholder="контактная информация" value="contact information" th:field="*{contactInformation}" required>
-        </div>
-        <div class="form-group mb-3 ">
-            <label for="creationDate">Дата создания</label>
-            <input type="date" class="form-control" id="creationDate" placeholder="01.01.2000" value="01.01.2001" th:field="*{creationDate}" required>
-        </div>
+    <h4><a href="/createWarehouse">Зарегистрировать склад</a></h4>
 
-        <button type="submit" class="btn btn-primary">Создать склад</button>
-        <a href="/static">Отменить</a>
-    </form>
-    <hr>
+    <tr>
+        <td colspan="7">
+            <form th:action="@{'/warehouses/' + ${warehouse.name}}" method="get">
+                <input type="text" name="filter" placeholder="Поиск по имени склада">
+                <button type="submit">Поиск</button>
+            </form>
+        </td>
+    </tr>
+
     <table>
         <thead>
         <tr>
@@ -58,14 +39,7 @@
         </tr>
         </thead>
         <tbody>
-        <tr>
-            <td colspan="7">
-                <form th:action="@{'/warehouses/' + ${warehouse.name}}" method="get">
-                    <input type="text" name="filter" placeholder="Поиск по имени склада">
-                    <button type="submit">Поиск</button>
-                </form>
-            </td>
-        </tr>
+
         <tr th:each="warehouse : ${warehouses}">
             <td th:text="${warehouse.name}"></td>
             <td th:text="${warehouse.address}"></td>
@@ -81,6 +55,9 @@
         </tr>
         </tbody>
     </table>
+
+    <h3><a href="/index">Вернутся в меню</a></h3>
+
 </div>
 
 <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css">
